Move parser definition to separate directory like other parsers, fix bug hidden by accidental scope inclusion of the wrong variable after that.
